create special color interpolation for fec objects.

+++ scilab/modules/renderer/src/java/org/scilab/modules/renderer/utils/geom3D/TriangleDecomposition.java
@@ -72,8 +72,8 @@ public class TriangleDecomposition {
 	 * @param color index of the color
 	 * @param polygonIndex of the polygon
 	 */
-	protected void setPolygonColor(int color, int polygonIndex) {
-		this.polygonColors[polygonIndex] = color;
+	protected void setPolygonColor(double color, int polygonIndex) {
+		this.polygonColors[polygonIndex] = (int) color;
 	}
